Reorganized file structure to make it more comprehensible and easy to navigate
