+++ app/components/App/styles.js
@@ -1,0 +1,17 @@
+import { StyleSheet } from 'react-native';
+export default StyleSheet.create({
+    container: {
+      flex: 1,
+      justifyContent: 'center',
+      alignItems: 'center',
+      backgroundColor: '#dfe6e9',
+    },
+    welcome: {
+      flex: 1,
+      fontSize: 20,
+      alignContent: 'flex-start',
+      justifyContent: 'flex-start',
+      margin: 15,
+      backgroundColor: '#dfe6e9',
+    },
+  });
