+++ app/components/Entry/Log-In/index.js
@@ -1,0 +1,2 @@
+import LoginForm from './LoginForm';
+export default LoginForm;
