+++ app/components/Entry/Log-In/styles.js
@@ -1,0 +1,28 @@
+import { StyleSheet } from 'react-native';
+export default StyleSheet.create({
+	containerStyle: {
+		padding: 20,	
+	},
+	inputStyle: {
+		height: 40,
+		backgroundColor: '#82ccdd',
+		marginBottom: 20,
+		paddingHorizontal: 10,
+	},
+	titleStyle: {
+		marginBottom: 20,
+		justifyContent: 'center',
+		alignContent: 'center',
+		textAlign: 'center',
+	},
+	buttonTextStyle: {
+		textAlign: 'center',
+		color: '#FFFFFF',
+		fontWeight: '700',
+	},
+	buttonContainerStyle: {
+		paddingVertical: 10,
+		backgroundColor: '#2980b9',
+		marginBottom: 10,
+	}
+});
