+++ app/components/Entry/index.js
@@ -1,0 +1,2 @@
+import Entry from './Entry';
+export default Entry;
