+++ app/components/Entry/styles.js
@@ -1,0 +1,28 @@
+import { StyleSheet } from 'react-native';
+export default StyleSheet.create({
+	containerStyle: {
+		flex: 1,
+		backgroundColor: '#60a3bc',		
+	},
+	logoContainerStyle: {
+		flexGrow: 1,
+		alignItems: 'center',
+		justifyContent: 'center',
+	},
+	logoStyle: {
+		width: 150,
+		height: 150,
+	},
+	titleStyle: {
+		color: '#2d3436',
+		marginTop: 10,
+	},
+	subtextStyle: {
+		color: '#2d3436',
+		marginTop: 10,
+		opacity: 0.6,
+	},
+	formContainerStyle: {
+        
+	}
+});
